+++ async_btree/__init__.py
@@ -14,15 +14,16 @@ from .decorator import (
     retry,
     retry_until_failed,
     retry_until_success,
+    ignore_exception,
 )
 from .definition import (
     FAILURE,
     SUCCESS,
     AsyncInnerFunction,
     CallableFunction,
-    ExceptionDecorator,
     NodeMetadata,
     node_metadata,
+    ControlFlowException,
 )
 from .leaf import action, condition
 from .parallele import parallele
@@ -41,6 +42,7 @@ __all__ = [
     'alias',
     'always_failure',
     'always_success',
+    'ignore_exception',
     'decorate',
     'inverter',
     'is_failure',
@@ -55,6 +57,7 @@ __all__ = [
     'ExceptionDecorator',
     'NodeMetadata',
     'node_metadata',
+    'ControlFlowException',
     'action',
     'condition',
     'parallele',