+++ tests/test_basics.py
@@ -1,6 +1,6 @@
 import pytest
 
-from async_btree import FAILURE, SUCCESS, ExceptionDecorator, node_metadata
+from async_btree import FAILURE, SUCCESS, ControlFlowException, node_metadata
 
 
 def test_truthy():
@@ -14,14 +14,21 @@ def test_falsy():
     assert not bool([])
     assert not FAILURE
     assert not bool(FAILURE)
-    assert not bool(ExceptionDecorator(Exception()))
+    assert not bool(ControlFlowException(Exception()))
 
 
 def test_exception_decorator_falsy():
     assert bool(Exception())
-    assert not bool(ExceptionDecorator(Exception()))
-    assert str(ExceptionDecorator(Exception("test"))) == str(Exception("test"))
-    assert repr(ExceptionDecorator(Exception("test"))) == repr(Exception("test"))
+    assert not bool(ControlFlowException(Exception()))
+    assert str(ControlFlowException(Exception("test"))) == str(Exception("test"))
+    assert repr(ControlFlowException(Exception("test"))) == repr(Exception("test"))
+
+
+def test_exception_deduplicate():
+    a = ControlFlowException(Exception("test 1"))
+    b = ControlFlowException(Exception("test 2"))
+    assert a != b
+    assert a == ControlFlowException.instanciate(a)
 
 
 @pytest.mark.curio