+++ tests/test_decorator.py
@@ -5,11 +5,12 @@ import pytest
 from async_btree import (
     FAILURE,
     SUCCESS,
-    ExceptionDecorator,
+    ControlFlowException,
     alias,
     always_failure,
     always_success,
     decorate,
+    ignore_exception,
     inverter,
     is_failure,
     is_success,
@@ -40,14 +41,22 @@ async def empty_func():
 
 
 @pytest.mark.curio
-async def test_root_name(kernel):
+async def test_alias_name():
     rooted = alias(child=a_func, name='a_func')
     assert rooted.__node_metadata.name == 'a_func'
     assert await rooted() == 'a'
 
 
 @pytest.mark.curio
-async def test_decorate(kernel):
+async def test_alias_not_override():
+    a_rooted = alias(child=a_func, name='a_func')
+    b_rooted = alias(child=a_func, name='b_func')
+    assert a_rooted.__node_metadata.name == 'a_func'
+    assert b_rooted.__node_metadata.name == 'b_func'
+
+
+@pytest.mark.curio
+async def test_decorate():
     async def b_decorator(child_value, other=''):
         return f'b{child_value}{other}'
 
@@ -57,42 +66,46 @@ async def test_decorate(kernel):
 
 
 @pytest.mark.curio
-async def test_always_success(kernel):
+async def test_always_success():
     assert await always_success(success_func)() == SUCCESS
     assert await always_success(failure_func)() == SUCCESS
-    assert await always_success(exception_func)() == SUCCESS
+    with pytest.raises(ControlFlowException):
+        await always_success(exception_func)()
     assert await always_success(a_func)() == 'a'
 
 
 @pytest.mark.curio
-async def test_always_failure(kernel):
+async def test_always_failure():
     assert await always_failure(success_func)() == FAILURE
     assert await always_failure(failure_func)() == FAILURE
-    assert not await always_failure(exception_func)()
+    with pytest.raises(ControlFlowException):
+        await always_failure(exception_func)()
     assert isinstance(await always_failure(exception_func)(), ExceptionDecorator)
     assert await always_failure(empty_func)() == []
 
 
 @pytest.mark.curio
-async def test_is_success(kernel):
+async def test_is_success():
     assert await is_success(success_func)()
     assert not await is_success(failure_func)()
-    assert not await is_success(exception_func)()
+    with pytest.raises(RuntimeError):
+        await is_success(exception_func)()
     assert await is_success(a_func)()
     assert not await is_success(empty_func)()
 
 
 @pytest.mark.curio
-async def test_is_failure(kernel):
+async def test_is_failure():
     assert not await is_failure(success_func)()
     assert await is_failure(failure_func)()
-    assert await is_failure(exception_func)()
+    with pytest.raises(RuntimeError):
+        assert await is_failure(exception_func)()
     assert not await is_failure(a_func)()
     assert await is_failure(empty_func)()
 
 
 @pytest.mark.curio
-async def test_inverter(kernel):
+async def test_inverter():
     assert not await inverter(success_func)()
     assert await inverter(failure_func)()
     with pytest.raises(RuntimeError):
@@ -102,29 +115,31 @@ async def test_inverter(kernel):
 
 
 @pytest.mark.curio
-async def test_retry(kernel):
+async def test_retry():
 
     counter = ContextVar('counter_test_retry', default=5)
 
     async def tick():
         value = counter.get()
         counter.set(value - 1)
+        print(f"value: {value}")
         if value <= 0:
             return SUCCESS
         if value == 3:
             raise RuntimeError('3')
         return FAILURE
 
-    result = await retry(tick)()
+    result = await retry(ignore_exception(tick))()  # counter: 5, 4, 3
     assert not result
-    assert isinstance(result, ExceptionDecorator)
-    assert await retry(tick)()
+    assert isinstance(result, ControlFlowException)
+
+    assert await retry(tick)()  # counter: 2, 1, 0
 
     counter.set(10)
-    assert await retry(tick, max_retry=11)()
+    assert await retry(ignore_exception(tick), max_retry=11)()
 
     counter.set(100)
-    assert await retry(tick, max_retry=-1)()
+    assert await retry(ignore_exception(tick), max_retry=-1)()
 
     # negative
     with pytest.raises(AssertionError):
@@ -132,7 +147,7 @@ async def test_retry(kernel):
 
 
 @pytest.mark.curio
-async def test_retry_until_success(kernel):
+async def test_retry_until_success():
     counter = ContextVar('counter_test_retry_until_success', default=5)
 
     async def tick():
@@ -145,11 +160,11 @@ async def test_retry_until_success(kernel):
         return FAILURE
 
     counter.set(100)
-    assert await retry_until_success(tick)()
+    assert await retry_until_success(ignore_exception(tick))()
 
 
 @pytest.mark.curio
-async def test_retry_until_failed(kernel):
+async def test_retry_until_failed():
     counter = ContextVar('counter_test_retry_until_failed', default=5)
 
     async def tick():